code reorg for fastpath file IO
