+++ Makefile.in
@@ -1,4 +1,4 @@
-px-objs = pxd.o dev.o iov_iter.o px_version.o io.o
+px-objs = pxd.o dev.o iov_iter.o px_version.o io.o kiolib.o
 obj-m = px.o
 
 KBUILD_CPPFLAGS := -D__KERNEL__
